Add fiber-store crate (#1168)
* add fiber-types crate

* remove store/schema.rs (moved to fiber-types)

* fix fiber-type import issues

* more refactor on fiber-types

* more refactor on types

* fix check-migration

* code cleanup

* move chain hash from fiber-types

* update ckb-* crates to 1

* fix fuzz

* code cleanup

* wasm32 compiling issue

* cleanup tentacle from fiber-types

* remove tentacle-secio from fiber-types

* fix ci check error

* code cleanup

* add cch feature gating for fiber-types

* fix default features for fiber-types

* seperate fiber-store from fiber-lib

* cleanup code

* remove  InMemorySignerExt from fiber-lib

* fix wasm test

* code review fixes
