+++ crates/fiber-lib/Cargo.toml
@@ -23,6 +23,7 @@ clap = { version = "4.5.2", features = ["derive", "env", "string"] }
 clap-serde-derive = "0.2.1"
 either = "1.15.0"
 fiber-sphinx = "2.3"
+fiber-store = { path = "../fiber-store" }
 fiber-types = { path = "../fiber-types", features = ["cch", "watchtower"] }
 futures = "0.3.30"
 getrandom = { version = "0.2", features = ["js"] }
@@ -57,15 +58,11 @@ biscuit-auth = "6.0.0-beta.3"
 ckb-hash = "1"
 clap = { version = "4.5.2", features = ["derive", "env", "string"] }
 clap-serde-derive = "0.2.1"
-console = "0.16.0"
 home = "0.5.9"
 hyper = { version = "1.5" }
 jsonrpsee = { version = "0.25.1", features = ["client", "server", "macros"] }
 lnd-grpc-tonic-client = "0.3.0"
 tonic = "0.11.0"
-rocksdb = { package = "ckb-rocksdb", version = "=0.21.1", features = [
-  "lz4",
-], default-features = false }
 tentacle = { version = "0.7", default-features = false, features = [
   "upnp",
   "parking_lot",
@@ -97,8 +94,6 @@ biscuit-auth = { version = "6.0.0-beta.3", features = ["wasm"] }
 ckb-hash = { version = "1", features = [
   "blake2b-ref",
 ], default-features = false }
-console_error_panic_hook = { version = "0.1.7" }
-fiber-wasm-db-common = { path = "../fiber-wasm-db-common" }
 getrandom = { version = "0.3", features = ["wasm_js"] }
 jsonrpsee = { version = "0.25.1", features = [
   "macros",
@@ -110,14 +105,12 @@ tentacle = { version = "0.7", default-features = false, features = [
 ] }
 tokio = { version = "1", features = ["io-util", "macros", "rt", "sync"] }
 tracing = { version = "0.1", features = ["log"] }
-wasm-bindgen = "0.2.84"
-web-sys = "0.3.72"
 web-time = { version = "1.1.0", features = ["serde"] }
 
 [features]
 bench = ["tempfile", "ckb-testtool"]
 default = ["watchtower"]
-portable = ["rocksdb/portable"]
+portable = ["fiber-store/portable"]
 sample = []
 watchtower = []
 metrics = ["dep:metrics", "ractor/metrics", "tentacle/metrics", "tokio-metrics", "metrics-exporter-prometheus"]
@@ -140,6 +133,7 @@ jsonrpsee = { version = "0.25.1", features = [
 ] }
 
 [target.'cfg(target_arch = "wasm32")'.dev-dependencies]
+fiber-store = { path = "../fiber-store", features = ["browser-test"] }
 wasm-bindgen-test = "0.3.50"
 
 [lints.clippy]
