+++ crates/fiber-lib/src/store/mod.rs
@@ -1,5 +1,3 @@
-pub mod db_migrate;
-pub mod migration;
 #[cfg(any(test, feature = "sample"))]
 pub mod sample;
 pub mod store_impl;
