+++ crates/fiber-lib/src/store/store_impl/native.rs
@@ -1,9 +1,8 @@
-use super::check_migrate;
 use super::{KVStore, KeyValue, StoreChange, StoreKeyValue};
+
+pub use fiber_store::DbDirection;
+pub use fiber_store::IteratorMode;
 use ractor::OutputPort;
-pub use rocksdb::Direction as DbDirection;
-pub use rocksdb::IteratorMode;
-use rocksdb::{prelude::*, DBCompressionType, WriteBatch, DB};
 use std::{fmt::Debug, path::Path, sync::Arc};
 
 pub trait StoreChangeWatcher: Send + Sync + Debug {
@@ -12,7 +11,7 @@ pub trait StoreChangeWatcher: Send + Sync + Debug {
 
 #[derive(Clone, Debug)]
 pub struct Store {
-    pub(crate) db: Arc<DB>,
+    pub(crate) inner: fiber_store::Store,
     watcher: Option<Arc<dyn StoreChangeWatcher>>,
 }
 
@@ -20,17 +19,16 @@ impl Store {
     /// Open a store, with migration check
     pub fn new<P: AsRef<Path>>(path: P) -> Result<Self, String> {
         let store = Self::open_db(path.as_ref())?;
-        let store = check_migrate(path, store)?;
+        let store = super::check_migrate(path, store)?;
         Ok(store)
     }
     /// Open a store, without migration check
     pub fn open_db(path: &Path) -> Result<Self, String> {
-        // add more migrations here
-        let mut options = Options::default();
-        options.create_if_missing(true);
-        options.set_compression_type(DBCompressionType::Lz4);
-        let db = Arc::new(DB::open(&options, path).map_err(|e| e.to_string())?);
-        Ok(Self { db, watcher: None })
+        let inner = fiber_store::Store::open_db(path)?;
+        Ok(Self {
+            inner,
+            watcher: None,
+        })
     }
 
     pub fn set_watcher(&mut self, watcher: Arc<dyn StoreChangeWatcher>) {
@@ -38,28 +36,26 @@ impl Store {
     }
 
     pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Option<Vec<u8>> {
-        self.db
-            .get(key.as_ref())
-            .map(|v| v.map(|vi| vi.to_vec()))
-            .expect("get should be OK")
+        self.inner.get(key)
     }
 
     pub(crate) fn delete<K: AsRef<[u8]>>(&self, key: K) {
-        self.db.delete(key).expect("Unexpected error from get");
+        self.inner.delete(key)
     }
 
     pub fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&self, key: K, value: V) {
-        self.db.put(key, value).expect("put should be ok");
+        self.inner.put(key, value)
     }
 
     pub fn batch(&self) -> Batch {
         Batch {
-            db: Arc::clone(&self.db),
-            wb: WriteBatch::default(),
+            inner: self.inner.batch(),
             watcher: self.watcher.clone().map(BatchWatcher::new),
         }
     }
-    /// Returns a prefix iterator, using iterator mode `mode`, skipping items until `skip_while` returns false, iterating over items prefixed with `prefix`
+
+    /// Returns a prefix iterator, using iterator mode `mode`, skipping items until `skip_while`
+    /// returns false, iterating over items prefixed with `prefix`
     #[allow(clippy::type_complexity)]
     pub fn prefix_iterator_with_skip_while_and_start<'a>(
         &'a self,
@@ -67,35 +63,22 @@ impl Store {
         mode: IteratorMode<'a>,
         skip_while: Box<dyn Fn(&[u8]) -> bool + 'static>,
     ) -> impl Iterator<Item = (Box<[u8]>, Box<[u8]>)> + 'a {
-        self.db
-            .get_iter(
-                &{
-                    let mut opts = ReadOptions::default();
-                    opts.set_prefix_same_as_start(true);
-                    opts
-                },
-                mode,
-            )
-            .skip_while(move |(key, _)| skip_while(key))
-            .take_while(move |(col_key, _)| col_key.starts_with(prefix))
+        self.inner
+            .prefix_iterator_with_skip_while_and_start(prefix, mode, skip_while)
     }
 
     pub fn prefix_iterator<'a>(
         &'a self,
         prefix: &'a [u8],
     ) -> impl Iterator<Item = (Box<[u8]>, Box<[u8]>)> + 'a {
-        self.prefix_iterator_with_skip_while_and_start(
-            prefix,
-            IteratorMode::From(prefix, DbDirection::Forward),
-            Box::new(|_| false),
-        )
+        self.inner.prefix_iterator(prefix)
     }
 }
 
 impl KVStore for Store {
-    /// Returns the underlying database instance for backup and maintenance.
-    fn inner_db(&self) -> &Arc<DB> {
-        &self.db
+    /// Returns the Store type.
+    fn inner_db(&self) -> &fiber_store::Store {
+        &self.inner
     }
 }
 
@@ -105,8 +88,7 @@ pub struct BatchWatcher {
 }
 
 pub struct Batch {
-    db: Arc<DB>,
-    wb: WriteBatch,
+    inner: fiber_store::Batch,
     watcher: Option<BatchWatcher>,
 }
 
@@ -151,29 +133,26 @@ impl BatchWatcher {
 
 impl Batch {
     pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Option<Vec<u8>> {
-        self.db
-            .get(key.as_ref())
-            .map(|v| v.map(|vi| vi.to_vec()))
-            .expect("get should be OK")
+        self.inner.get(key)
     }
 
     pub fn put_kv(&mut self, key_value: KeyValue) {
-        self.put(key_value.key(), key_value.value());
+        self.inner.put(key_value.key(), key_value.value());
         if let Some(watcher) = &mut self.watcher {
             watcher.record_put(key_value);
         }
     }
 
     pub fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&mut self, key: K, value: V) {
-        self.wb.put(key, value).expect("put should be OK")
+        self.inner.put(key, value)
     }
 
     pub fn delete<K: AsRef<[u8]>>(&mut self, key: K) {
-        self.wb.delete(key.as_ref()).expect("delete should be OK");
+        self.inner.delete(key)
     }
 
     pub fn commit(self) {
-        self.db.write(&self.wb).expect("commit should be OK");
+        self.inner.commit();
         if let Some(watcher) = self.watcher {
             watcher.commit();
         }
