+++ crates/fiber-store/src/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum StoreError {
+    #[error("Database internal error: {0}")]
+    DBInternalError(String),
+}
